### 2019-05-18 #5 @marskar
Rmdawn: (de)construct, convert, and render R Markdown (IN SCOPE) 
Submitting Author: Name (@marskar)
Repository Link: https://github.com/marskar/rmdawn
Documentation Link: https://marskar.github.io/rmdawn/
PyPI Link: https://pypi.org/project/rmdawn/

---

```
The Rmdawn Python package allows you to
- (de)construct,
- convert, and
- render
R Markdown (Rmd) files.

The Rmdawn Python package consists of 6 shell commands and functions:

- rmdawn, which concatenates input files to generate an Rmd file.
- rmdusk, which extracts
    - a YAML file,
    - Python or R scripts and
    - Markdown (md) files from Rmd files.
- rmdtor, which converts Rmd files into R scripts using knitr::purl.
- rtormd, which converts R scripts into Rmd files using knitr::spin.
- render, which creates rendered versions of R scripts or Rmd files into
    - HTML,
    - PDF,
    - Word, and
    - other output file formats.
- catren, which combines the functionality of
    - rmdawn and
    - render to generate an Rmd file from source files and then create an output file.

All Rmdawn functions and commands,
except for rmdawn and rmdusk, rely on the rpy2 Python library.
The command line interface relies on the click Python library.
```

- Please indicate which [category or categories](https://pyopensci.github.io/dev_guide/peer_review/peer_review_proc.html#aims-and-scope) this package falls under:

R markdown is important for reproducibility, because it facilitates literate programming, the combination of code and text. This way, descriptions of methods and results can be adjacent to the code used to obtain the results.
I frequently use R markdown for my work and I want to make it easier for the students in the Python classes I teach to use R markdown. With Rmdawn, anyone can create R markdown files from simple source files, or extract source files from R markdown files. This can promote the modularity principle without sacrificing the benefits of literate programming or the convenience of a single file with all code, text, and output for the project.

The target audience is anyone who uses R markdown. This package may help teachers and researchers to improve the efficiency of their R markdown workflows. The package aims to be useful for both R users and non-users. The `rmdawn` and `rmdusk` do not even require R to be installed! I imagine teachers and researchers writing markdown and code files and then using `rmdawn` to creating R markdown files to share with their students and colleagues.

Related packages include:

- [`notedown`](https://github.com/aaren/notedown)

Notedown can all convert Jupyter notebooks into markdown and R markdown files. Rmdawn is different, because it can take a more modular approach to file conversion by extracting the contents of each code chunk into a separate file (chunk -> file) or using a source file to create each code chunk (file -> chunk).

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*
