### 2019-10-28 #5 @mbjoseph
@marskar this looks like this would be in-scope for us! I will be the handling editor for Rmdawn.

I'll close this presubmission issue, and you can open a new submission issue using the "Submit Software for Review" template whenever you are ready. Let me know if you have any questions!
