### 2019-11-13 #5 @lwasser
hey @marskar just checking in here -- are you still interested in us reviewing rmdawn? if so, please open a new software submission issue here and we can get the review process going with @mbjoseph  as editor.
